# Improve path completion