… src/System.Management.Automation/utils/PowerShellExecutionHelper.cs
… test/powershell/Host/TabCompletion/TabCompletion.Tests.ps1