@bradwilson 2023-11-22
> Also stopped it from replacing ~ when referring to home so this: ~\Downloads<Tab> no longer turns into: C:\Users\Martin\Downloads\

This is an unfortunate change. It's broken me a lot because those tildes not being replaced means that applications no longer understand paths with it in it. I got a suggestion to use `Enable-ExperimentalFeature PSNativePSPathResolution` but that is apparently no longer an option.

Is there a way for me to turn this back on? I'm getting super tired of having to retype `C:\Users\bradwilson`. 😢